Project/requirements.txt
Flask
beautifulsoup4
datetime
lxml
playwright
requests
firebase-admin
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client